refactor: using PascalCase filenames for react components
