+++ app/components/pages/SubmissionWizard/ProgressBar.js
