+++ app/components/pages/SubmissionWizard/steps/Author/AuthorDetails.md
