+++ app/components/pages/SubmissionWizard/steps/Author/AuthorDetails.test.js
