+++ app/components/pages/SubmissionWizard/steps/Author/index.js
