+++ app/components/pages/SubmissionWizard/steps/Author/schema.js
