+++ app/components/pages/SubmissionWizard/steps/Editors/FormSections.js
