+++ app/components/pages/SubmissionWizard/steps/Editors/index.js
