+++ app/components/pages/SubmissionWizard/steps/Editors/index.test.js
