+++ app/components/pages/SubmissionWizard/steps/Editors/schema.js
