+++ app/components/pages/SubmissionWizard/steps/Files/index.js
