+++ app/components/pages/SubmissionWizard/steps/Files/schema.js
