+++ app/components/pages/SubmissionWizard/steps/Submission/ManuscriptMetadata.md
